get_gdeltnews

Get last 2 weeks of GDELT COVID news. Deduplicated by DateTime, URL, and DomainCountryCode.

Usage

get_gdeltnews(period = 14, start_date = Sys.Date() - 1)

Arguments

period

(numeric) Number of days to look back from start_date

start_date

(Date) Date from which to look back

Value

Returns a data frame with n rows and 16 columns, including:

DateTime POSIXct

URL character

Title character

SharingImage character

LangCode character

DocTone double

DomainCountryCode character

Location character

Lat double

Lon double

iso2code character

Adm1Code character

GeoType character

ContextualText character

GeoCoord character

Adm2Code character

Note

Function may fail if start date is set to current date or in future. News data files aren't often avilable for the current date.

Examples

if (FALSE) { # \dontrun{
gdeltnews_df <- get_gdeltnews()
} # }