Our World in Data (OWID) Testing Metadata

A metadata table used within get_testing() to determine how country testing data is computed.

To scrape the OWID coronavirus testing website for data related to the source and definitions of their testing data by country. Note that OWID testing data will no longer be updated from 23 June 2022 on.

Usage

owid_testing_meta

get_owid_testing_meta(url = "https://ourworldindata.org/coronavirus-testing")

Format

A data frame with 141 rows and 5 columns, including:

id character ISO 3166-1 alpha-3 country code

test_definition character Description of OWID testing definition

case_definition character Description of OWID case definition

posrate_definition character Description of OWID test positivity calculation

posrate_direct logical Indicator for whether OWID pulls test positivity directly, or computes it

Value

Data set of scraped information related to data definitions and sources by country

Functions

owid_testing_meta: Saved dataset

Examples

if (FALSE) { # \dontrun{
owid_testing_meta <- get_owid_testing_meta()
usethis::use_data(owid_testing_meta, overwrite = T)
} # }