Get testing data from Our World in Data (OWID) or FIND

Pulls testing data from either OWID or FIND sources based on a 7d testing / 1K and positivity rate.

Most recent 7-day average data points for new tests per 1K AND positivity rate for use in risk matrix plots by WHO region.

After accounting for flags, we pick OWID by default when both OWID and FIND are available in last X days

Usage

get_testing(analysis_date = Sys.Date() - 1L)

Arguments

analysis_date

(date default: Sys.Date() - 1L) Start date from which to choose OWID or FIND data

Value

Returns a data frame with n rows and 5 columns, including:

id

character ISO 3166-1 alpha-3 country code

date

date Date of testing observation

new_tests_smoothed_per_thousand

double Tests / 1000 population (7d smooth)

new_tests_smoothed_per_thousand_14

double Tests / 1000 population (14d smooth)

positive_rate

double Test positivity rate (7d smooth)

Examples

if (FALSE) { # \dontrun{
testing <- get_testing()
} # }