Get data on count of vaccinations by vaccine manufacturer from OWID

Returns number of vaccination doses delivered by manufacturer. Note that data are not a daily time series but are reported intermittently by countries. Data are pulled in from OWID. Only a subset of countries have this data available, and information on which countries and the associated source of data can be found here.

Usage

get_vax_by_type()

Value

Returns a data frame with n rows and 6 columns, including:

owid_country

character, English country name from OWID (may not match WHO country name)

id

character, ISO 3166-1 alpha-3 country code

date

date, Date of observation

vaccine

character, vaccine manufacturer

total_vaccinations

integer, cumulative number of vaccinations for a given date

new_vaccinations

integer, new vaccinations for a given date interval, note that this is not actually daily, but depends on the cadence of reporting