Compute Percent-change in Cases Over a Specified Window

Given a time series of new cases over a certain date range, compute a windowed percent change value of cases (i.e. one week versus previous 7 days, etc.).

The window is right-aligned by the date column (beginning the most recent date shifting backwards).

Note that because we're computing based on index rather than calendar time, results will be erroneous if data are not complete for every date.

Usage

calc_window_pct_change(df, window = 14L, return_totals = FALSE)

Arguments

df

Dataframe with at least date, new_cases columns

window

(default: 14) number of days the comparison windows should be

return_totals

(default: FALSE) return running sums used to compute pct_change?

Value

a df summarized by date with new column pct_change, or [pct_change, cases_current_window, cases_prev_window] if return_totals is TRUE

Examples

if (FALSE) {
data <- get_covid_df()
calc_window_pct_change(data, window = 14)
# For grouped operations, group data beforehand and pipe:
data |>
  group_by(iso2code, country) |>
  calc_window_pct_change(data, window = 14)
}