get_vax_manufacturers

Get vaccination metadata from OWID

Usage

get_vax_manufacturers()

Value

A data frame with 218(-ish) rows and 6 variables:

owid_country character English country name from OWID

id character ISO 3166-1 alpha-3 country code

vaccines character list of vaccines administered in the country up to the current date.

last_observation_date date date of the last observation in OWID data.

source_name character name of OWID source for data collection.

source_website character web location of OWID source. It can be a standard URL if numbers are consistently reported on a given page; otherwise it will be the source for the last data point.